+++ M2XStreamClient/m2x-esp8266.h
@@ -15,5 +15,5 @@ class M2XTimer {
 public:
   void start() {}
 
-  int read_ms() { return millis(); }
+  unsigned long read_ms() { return millis(); }
 };
